requirements.txt
flask==2.3.2
pymysql==1.0.2
scikit-learn==1.2.2
gunicorn==20.1.0
numpy==2.3.1
pandas==2.3.0
plotly==6.2.0
mysql-connector-python==9.3.0
PyMySQL==1.0.2
